It's been 4 years...
... and I release a sigh of relief and sadness. 

I'm relieved that we didn't mess it up. We built the most complicated
robot, in my opinion, that 610's ever built. Turret with android
tracking, gear manipulator on a slider, ball intake, and a complicated
hopper all packed into the smallest volume ever allowed in the FRC
rule-book. Although it's not the best robot in FRC and it has its flaws,
it will sit as #1 in my heart.

I'm sad that this is the final build season of my life. I will never
have the same experience working with a group of guys building such a
fine machine in only six weeks. Wherever I end up, I'll always remember
the sick J's and the dramas.

There's a long road ahead of us, and I will continue to persevere into
the competition season until we get that Einstein gold again.

Changes:
- T_SmartDash for easier debugging
- Updated all tuned PID Constants
- Adaptive feeder control
- PDB code yet to test
- Cleaned up T_Shooter

- My life